requirements.txt

NumPy
Requests
streamlit==1.40.0
pandas==2.2.2
yfinance==0.2.43
plotly==6.4.0
PyYAML==6.0.2
bcrypt==4.1.2
streamlit-authenticator==0.3.2
numpy==1.26.5
matplotlib==3.8.4
